requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==1.4.51
python-multipart==0.0.9
jinja2==3.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1 